@@ -1,10 +1,10 @@
 {
   "name": "bilibilivideodownload",
-  "version": "3.2.0",
+  "version": "3.3.0",
   "private": true,
   "author": "wangyu <wangyu@wangyu.link>",
   "description": "欢迎使用BilibiliVideoDownload",
-  "homepage": "blog.wangyu.link",
+  "homepage": "https://github.com/blogwy/BilibiliVideoDownload",
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
@@ -14,38 +14,45 @@
     "postinstall": "electron-builder install-app-deps",
     "postuninstall": "electron-builder install-app-deps"
   },
-  "main": "background.js",
   "dependencies": {
+    "@ant-design/icons-vue": "^6.1.0",
     "@ffmpeg-installer/ffmpeg": "^1.1.0",
-    "ant-design-vue": "^1.7.8",
-    "core-js": "^3.6.5",
+    "@types/lodash": "^4.14.182",
+    "@types/qrcode": "^1.4.2",
+    "ant-design-vue": "^3.2.2",
+    "dayjs": "^1.11.1",
+    "electron-log": "^4.4.6",
     "electron-store": "^8.0.1",
+    "fluent-ffmpeg": "^2.1.2",
+    "fs-extra": "^10.1.0",
     "got": "^11.8.3",
+    "lodash": "^4.17.21",
+    "nanoid": "^3.3.3",
+    "pinia": "^2.0.13",
     "qrcode": "^1.5.0",
-    "vue": "^2.6.11",
-    "vue-perfect-scrollbar": "^0.2.1",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "vue": "^3.2.13",
+    "vue-router": "^4.0.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-standard": "^5.1.2",
-    "babel-eslint": "^10.1.0",
-    "electron": "^13.0.0",
+    "@types/electron-devtools-installer": "^2.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-plugin-router": "~5.0.0",
+    "@vue/cli-plugin-typescript": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "@vue/eslint-config-standard": "^6.1.0",
+    "@vue/eslint-config-typescript": "^9.1.0",
+    "electron": "^16.0.0",
     "electron-devtools-installer": "^3.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
+    "eslint": "^7.32.0",
+    "eslint-plugin-import": "^2.25.3",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "less": "^3.0.4",
-    "less-loader": "^5.0.0",
-    "vue-cli-plugin-electron-builder": "~2.1.1",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-vue": "^8.0.3",
+    "less": "^4.0.0",
+    "less-loader": "^8.0.0",
+    "typescript": "~4.5.5",
+    "vue-cli-plugin-electron-builder": "^3.0.0-alpha.4"
   }
 }
